getting ready for PR and 0.0.2 release
