+++ asyncbolt/client.py
@@ -24,7 +24,7 @@ async def connect(*,
                   port=None,
                   ssl=None,
                   on_failure=None,
-                  max_inflight=1024,
+                  max_inflight=2048,
                   **kwargs):
     """
     Connect to the Bolt server and initialize session
@@ -158,6 +158,8 @@ class ClientSession:
         self._inflight += 2
 
     async def reset(self):
+        """???What is expected client session behavior on RESET???"""
+        # TODO investigate this method
         if self._inflight > self._max_inflight:
             raise BoltClientError('Exceeded max number of pipelined messages')
         self._protocol.reset()
