+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup
 
 setup(
     name="asyncbolt",
-    version="0.0.1",
+    version="0.0.2",
     license="MIT",
     author="David M. Brown",
     author_email="davebshow@gmail.com",
